bug_fix double to int

+++ lib/widgets/new_transaction.dart
@@ -21,7 +21,7 @@ class _NewTransactionState extends State<NewTransaction> {
       return;
     }
     final enteredTitle = _titleController.text;
-    final enteredAmount = double.parse(_amountController.text);
+    final enteredAmount = int.parse(_amountController.text);
 
     if (enteredTitle.isEmpty || enteredAmount <= 0 || _selectedDate == null) {
       return;
